feat: test positions
